Refactor Neuronenblitz package and add config sync service
