+++ FAILEDTESTS.md
@@ -1,1 +1,3 @@
-All tests pass as of Mon Jul 28 06:43:30 UTC 2025.
+Failed tests after latest changes:
+- tests/test_remote_offloading.py::test_bandwidth_and_route
+- tests/test_streamlit_gui.py::test_function_search_count_synapses (timeout)
