+++ marble_neuronenblitz/__init__.py
@@ -1,0 +1,15 @@
+from .core import (
+    Neuronenblitz,
+    default_combine_fn,
+    default_loss_fn,
+    default_weight_update_fn,
+    default_q_encoding,
+)
+
+__all__ = [
+    "Neuronenblitz",
+    "default_combine_fn",
+    "default_loss_fn",
+    "default_weight_update_fn",
+    "default_q_encoding",
+]
